fix for external
